requirements.txt
torch
transformers
Pillow
numpy
scipy